Light Edits Post Audio
Edited break script to have a stronger check for tag switching so we don't have our audio for switching playing on loop (ow my ears). Made final edits to the positions of hands so that rotations feel nice and the player feels that their hands are attached to the game's hands.